dateutil  monthNameShort & getDaysinMoth now work with array instaed of big switch #refactor

+++ src/Utils/DateUtil.h
@@ -9,7 +9,7 @@ class DateUtil {
 	public:
 
 		static uint8_t weekday(uint year, uint8_t month, uint8_t day);
-		static uint8_t getDaysInMonth(uint year, uint8_t month);
-		static void monthNameShort(char (&name)[4], uint8_t month);
+		static uint8_t getDaysInMonth(uint year, int month);
+		static void monthNameShort(char (&name)[4], int month);
 
 };
